+++ src/kubernetes/api/rbac_authorization_k8s_io_v1.clj
@@ -1,7 +1,7 @@
-(ns kubernetes.api.security-istio-io-v1
+(ns kubernetes.api.rbac-authorization-k8s-io-v1
   (:require [kubernetes.api.openapiv3 :as openapiv3]
             [kubernetes.api.util :as util]))
 
 (def make-context util/make-context)
 
-(openapiv3/render-full-api "security.istio.io_v1")
+(openapiv3/render-full-api "rbac.authorization.k8s.io_v1")