+++ src/kubernetes/api/telemetry_istio_io_v1.clj
@@ -1,7 +1,7 @@
-(ns kubernetes.api.security-istio-io-v1
+(ns kubernetes.api.telemetry-istio-io-v1
   (:require [kubernetes.api.openapiv3 :as openapiv3]
             [kubernetes.api.util :as util]))
 
 (def make-context util/make-context)
 
-(openapiv3/render-full-api "security.istio.io_v1")
+(openapiv3/render-full-api "telemetry.istio.io_v1")